+++ src/include/skalibs/bufalloc.h
@@ -3,6 +3,7 @@
 #ifndef BUFALLOC_H
 #define BUFALLOC_H
 
+#include <sys/types.h>
 #include <skalibs/gccattributes.h>
 #include <skalibs/stralloc.h>
 
@@ -10,14 +11,14 @@ typedef struct bufalloc bufalloc, *bufalloc_ref ;
 struct bufalloc
 {
   stralloc x ;
-  unsigned int p ;
+  size_t p ;
   int fd ;
   int (*op) (int, char const *, unsigned int) ;
 } ;
 
 #define BUFALLOC_ZERO { .x = STRALLOC_ZERO, .p = 0, .fd = -1, .op = 0 }
 #define BUFALLOC_INIT(f, d) { .x = STRALLOC_ZERO, .p = 0, .fd = (d), .op = (f) }
-extern void bufalloc_init (bufalloc *, int (*)(int, char const *, unsigned int), int) ;
+extern void bufalloc_init (bufalloc *, int (*)(int, char const *, size_t), int) ;
 #define bufalloc_shrink(ba) stralloc_shrink(&(ba)->x)
 #define bufalloc_free(ba) stralloc_free(&(ba)->x)
 #define bufalloc_put(ba, s, n) stralloc_catb(&(ba)->x, s, n)
@@ -28,7 +29,7 @@ extern int bufalloc_getfd (bufalloc const *) gccattr_pure ;
 extern int bufalloc_flush (bufalloc *) ;
 extern void bufalloc_clean (bufalloc *) ;
 #define bufalloc_len(ba) ((ba)->x.len - (ba)->p)
-extern unsigned int bufalloc_getlen (bufalloc const *) gccattr_pure ;
+extern size_t bufalloc_getlen (bufalloc const *) gccattr_pure ;
 #define bufalloc_isempty(ba) ((ba)->x.len == (ba)->p)
 
 extern bufalloc bufalloc_1_ ;
