+++ src/include/skalibs/djbunix.h
@@ -35,7 +35,7 @@ extern int fd_chmod (int, unsigned int) ;
 extern int fd_chown (int, unsigned int, unsigned int) ;
 extern int fd_sync (int) ;
 extern int fd_cat (int, int) ;
-extern unsigned int fd_catn (int, int, unsigned int) ;
+extern size_t fd_catn (int, int, size_t) ;
 extern int fd_ensure_open (int, int) ;
 #define fd_sanitize() (fd_ensure_open(0, 0) && fd_ensure_open(1, 1) && fd_ensure_open(2, 1))
 extern int lock_ex (int) ;
