+++ src/libstddjb/bitarray_firstclear_skip.c
@@ -1,11 +1,12 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bitarray.h>
 
-unsigned int bitarray_firstclear_skip (register unsigned char const *s, unsigned int max, unsigned int skip)
+size_t bitarray_firstclear_skip (register unsigned char const *s, size_t max, size_t skip)
 {
-  register unsigned int i = bitarray_div8(skip) ;
-  register unsigned int j = i << 3 > max ? max : i << 3 ;
+  register size_t i = bitarray_div8(skip) ;
+  register size_t j = i << 3 > max ? max : i << 3 ;
   if (i && s[i-1] != 0xffU)
   {
     while ((skip < j) && bitarray_peek(s, skip)) skip++ ;
