+++ src/libstddjb/bitarray_not.c
@@ -1,8 +1,9 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bitarray.h>
 
-void bitarray_not (register unsigned char *s, register unsigned int a, register unsigned int b)
+void bitarray_not (register unsigned char *s, register size_t a, register size_t b)
 {
   if (!b) return ;
   b += a ;
@@ -10,7 +11,7 @@ void bitarray_not (register unsigned char *s, register unsigned int a, register 
     s[a>>3] ^= ((1 << (a & 7)) - 1) ^ ((a << (b & 7)) - 1) ;
   else
   {
-    register unsigned int i = (a>>3) + 1 ;
+    register size_t i = (a>>3) + 1 ;
     s[a>>3] ^= ~((1 << (a & 7)) - 1) ;
     for (; i < (b>>3) ; i++) s[i] = ~s[i] ;
     s[b>>3] ^= (1 << (b & 7)) - 1 ;
