+++ src/libstddjb/fd_cat.c
@@ -1,16 +1,17 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/iobuffer.h>
 #include <skalibs/djbunix.h>
 
 int fd_cat (int from, int to)
 {
   iobuffer b ;
-  unsigned int n = 0 ;
+  size_t n = 0 ;
   if (!iobuffer_init(&b, from, to)) return -1 ;
   for (;;)
   {
-    register int r = iobuffer_fill(&b) ;
+    register ssize_t r = iobuffer_fill(&b) ;
     if (r < 0) goto err ;
     else if (!r) break ;
     if (!iobuffer_flush(&b)) goto err ;
