+++ src/libstddjb/fd_recv.c
@@ -5,9 +5,9 @@
 #include <errno.h>
 #include <skalibs/allreadwrite.h>
 
-int fd_recv (int fd, char *buf, unsigned int len, unsigned int flags)
+ssize_t fd_recv (int fd, char *buf, size_t len, unsigned int flags)
 {
-  register int r ;
+  register ssize_t r ;
   do r = recv(fd, buf, len, (int)flags) ;
   while ((r == -1) && (errno == EINTR)) ;
   return r ;
