+++ src/libstddjb/genwrite_put_bufalloc.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bufalloc.h>
 #include <skalibs/genwrite.h>
 
-int genwrite_put_bufalloc (void *target, char const *s, unsigned int len)
+ssize_t genwrite_put_bufalloc (void *target, char const *s, size_t len)
 {
   register bufalloc *ba = target ;
-  return bufalloc_put(ba, s, len) ? (int)len : -1 ;
+  return bufalloc_put(ba, s, len) ? (ssize_t)len : -1 ;
 }
